fix(widgets): address PR review feedback
- Fix HTMLElement tooltip never refreshing (use replaceChildren)
- Fix ThemeWidget tooltip using wrong mode branch
- Add tooltip tests: string override, false disables, HTMLElement
  content, HTMLElement replacement
- Add ThemeWidget tooltip tests (mode alignment, label fallback)
- Document tooltip override props on all individual widget API pages
- Document automatic IconButton tooltip behavior for Selector, Loading,
  Timeline, Stats, and Popup widgets
